# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TEST TUL API Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "juan.cuero@unillanos.edu.co"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/api/v1/cart/show": {
            "get": {
                "tags": [
                    "Cart"
                ],
                "summary": "Show cart",
                "description": "<b>Returns the current cart.</b> <br> \n                       Creation Date: 13/04/2021 05:30 PM <br> \n                       Create By: Juan Cuero <br>\n                    Last Edit Date: 13/04/2021 05:30 PM <br> \n            ",
                "operationId": "App\\Http\\Controllers\\v1\\CartController::show",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/cart/{product}/add": {
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Add product to cart",
                "description": "<b>Agregar Producto a carrito.</b> <br> \n                       Creation Date: 13/04/2021 05:00 PM <br> \n                       Create By: Juan Cuero <br>\n                    Last Edit Date: 13/04/2021 05:00 PM <br> \n    *        ",
                "operationId": "App\\Http\\Controllers\\v1\\CartController::store",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "quantity": {
                                        "title": "Quantity",
                                        "type": "integer",
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/cart/{product}/remove": {
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "remove product of the cart",
                "description": "<b>Remover Producto de carrito.</b> <br> \n                       Creation Date: 14/04/2021 04:20 PM <br> \n                       Create By: Juan Cuero <br>\n                    Last Edit Date: 14/04/2021 04:20 PM <br> \n    *        ",
                "operationId": "App\\Http\\Controllers\\v1\\CartController::remove",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "quantity": {
                                        "title": "Quantity",
                                        "type": "integer",
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/cart/confirm": {
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Confirm cart",
                "description": "<b>Confirm cart.</b> <br> \n     *                 Creation Date: 14/04/2021 05:20 PM <br> \n     *                 Create By: Juan Cuero <br>\n     *              Last Edit Date: 14/04/2021 05:20 PM <br> \n     *        ",
                "operationId": "App\\Http\\Controllers\\v1\\CartController::confirm",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/categories": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "List categories",
                "description": "<b>Returns the list of all categories.</b> <br> \n                       Creation Date: 14/04/2021 08:20 PM <br> \n                       Create By: Juan Cuero <br>\n                    Last Edit Date: 15/04/2021 03:20 PM <br> \n            ",
                "operationId": "App\\Http\\Controllers\\v1\\CategoryController::index",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "Category new",
                "description": "<b>Create new category.</b> <br> \n                       Creation Date: 15/04/2021 03:45 PM <br> \n                       Create By: Juan Cuero <br>\n                    Last Edit Date: 15/04/2021 03:45 PM <br> \n    *        ",
                "operationId": "App\\Http\\Controllers\\v1\\CategoryController::store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CategoryCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/categories/{category}/changestatus": {
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "Change status category",
                "description": "<b>Change status categor.</b> <br> \n                       Creation Date: 14/04/2021 06:00 PM <br> \n                       Create By: Juan Cuero <br>\n                    Last Edit Date: 14/04/2021 06:00 PM <br> \n    *        ",
                "operationId": "App\\Http\\Controllers\\v1\\CategoryController::changeStatus",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "category id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "List Products",
                "description": "<b>Returns the list of available products.</b> <br> \n                       Creation Date: 13/04/2021 04:00 PM <br> \n                       Create By: Juan Cuero <br>\n                    Last Edit Date: 13/04/2021 04:00 PM <br> \n            ",
                "operationId": "App\\Http\\Controllers\\v1\\ProductController::index",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Crear producto",
                "description": "<b> Crea producto. </b> <br> \n      *                  Creation Date: 14/04/2021 07:00 PM<br> \n     *                   Create By: Juan Cuero <br>\n     *                   Last Edit Date: 14/04/2021 07:00 PM <br> \n    *       ",
                "operationId": "App\\Http\\Controllers\\v1\\ProductController::store",
                "requestBody": {
                    "description": "Bulk products Body",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "These credentials do not match our records."
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "CategoryCreate": {
                "description": "<b> Register Category model</b> <br>",
                "properties": {
                    "name": {
                        "title": "Name",
                        "description": "Name of the category",
                        "type": "string",
                        "example": "Materiales de Construcción"
                    },
                    "parent_id": {
                        "title": "Parent",
                        "description": "Parent of the category",
                        "type": "integer",
                        "example": null
                    }
                },
                "type": "object"
            },
            "ProductCreate": {
                "description": "<b> Register Product model</b> <br>",
                "properties": {
                    "name": {
                        "title": "Name",
                        "description": "Name of the product",
                        "type": "string",
                        "example": "Mi producto"
                    },
                    "description": {
                        "title": "Description",
                        "description": "Description of the product",
                        "type": "string",
                        "example": "Hola esta es mi primer producto"
                    },
                    "stock": {
                        "title": "Stock",
                        "description": "stock of the product",
                        "type": "integer",
                        "example": 20
                    },
                    "price": {
                        "title": "Price",
                        "description": "price of the product",
                        "type": "decimal",
                        "example": 50000
                    },
                    "category_id": {
                        "title": "Category",
                        "description": "Category of the product",
                        "type": "integer",
                        "example": 3
                    },
                    "image": {
                        "title": "Image",
                        "description": "Image of the product",
                        "type": "string",
                        "format": "binary"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Products",
            "description": "API Endpoints of Products"
        },
        {
            "name": "Cart",
            "description": "API Endpoints of Cart"
        }
    ],
    "security": [
        []
    ]
}